more float to CGFloat
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@3940 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/Addresses/Frameworks/AddressView/ADPersonPropertyView.h
@@ -141,11 +141,11 @@ typedef enum {
 	   changedToValue: (id) value
 		   inView: (id) sender;
 - (void) view: (id) view
-changedWidthFrom: (float) oldW
-	   to: (float) newW;
+changedWidthFrom: (CGFloat) oldW
+	   to: (CGFloat) newW;
 - (void) view: (id) view
-changedHeightFrom: (float) oldH
-	   to: (float) newH;
+changedHeightFrom: (CGFloat) oldH
+	   to: (CGFloat) newH;
 - (void) beginEditingInNextViewWithTextMovement: (int) textMovement;
 
 - (BOOL) personPropertyView: (ADPersonPropertyView*) view
